# [refactor] - Improve Performance by Shifting Concurrency from Image to Layer Level