<!--
Please create an issue to collect feedback prior to feature additions. Please also reference that issue in any PRs.
If possible try to keep PRs scoped to one feature, and add tests for new features.
-->

### Description:
This PR enhances the scanning performance and resource utilization of our Docker image scanning process by moving the concurrency logic from the image level to the layer level.

*Details:*
- Previously, our scanning process handled one image at a time, processing its layers sequentially. This approach limited the efficiency and speed, especially with multi-core systems, as only one core could be utilized at a time.
- This update introduces concurrent scanning of individual layers within each image. By parallelizing the scanning of layers, we can significantly improve performance by taking advantage of available CPU cores.
- The new implementation allows multiple layers to be scanned simultaneously, leading to faster processing times and better resource utilization.

#### Benchmarks
**BEFORE**:
- MySQL: 10s
- Postgres: 6s
- Ubuntu: 2.5s
- Python: 11s

**AFTER**:
- MySQL: 7s
- Postgres: 4s 
- Ubuntu: 1.8s
- Python: 7.5s

### Checklist:
* [ ] Tests passing (`make test-community`)?
* [ ] Lint passing (`make lint` this requires [golangci-lint](https://golangci-lint.run/usage/install/#local-installation))?